--- applications/luci-app-macvlan/po/zh-cn/macvlan.po
@@ -1,52 +1,0 @@
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2014-06-28\n"
-"PO-Revision-Date: 2014-06-28\n"
-"Last-Translator: 981213 <gch981213@gmail.com>\n"
-"Language: zh_CN\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-msgid "Virtual WAN"
-msgstr "虚拟WAN接口"
-
-msgid "Create virtual WAN interfaces"
-msgstr "创建虚拟WAN接口"
-
-msgid "Here you can create some virtual WAN interfaces with MACVLAN driver."
-msgstr "在这里，你可以使用Macvlan驱动创建虚拟WAN接口。"
-
-msgid "Here you can create some virtual WAN interfaces with MACVLAN driver.Set wan metric first!<br />Connected interface count:"
-msgstr "在这里，你可以使用Macvlan驱动创建虚拟WAN接口。请务必先设置WAN口的网关跃点（通常可以填写 40 ）。<br />已连接的WAN接口数量： "
-
-msgid "Enable"
-msgstr "启用"
-
-msgid "rediag"
-msgstr "重新并发拨号"
-
-msgid "Number of virtual WAN"
-msgstr "虚拟WAN接口数量"
-
-msgid "Number of PPP diag"
-msgstr "PPPoE并发拨号数量"
-
-msgid "Use WAN to diag"
-msgstr "使用WAN口进行拨号"
-
-msgid "Uncheck this if you use a VLAN interface."
-msgstr "如果你使用的是VLAN接口（WAN对应的接口是ethX.X）,请取消勾选这个选项。"
-
-msgid "Enable auto reconnect"
-msgstr "断线自动重连"
-
-msgid "Minimnum connected interface"
-msgstr "最低在线接口数"
-
-msgid "Rediag if the number of connected interfaces is less then this."
-msgstr "如果在线接口数低于这个数值，将会自动重新拨号。"
-
