--- applications/luci-app-macvlan/root/etc/config/macvlan
@@ -1,6 +1,0 @@
-config macvlan 'config'
-	option 'enabled'	'0'
-	option 'wannum'		'5'
-	option 'wansw'		'0'
-	option 'diagchk'	'0'
-	option 'diagnum'	'1'
