+++ applications/luci-app-syncdial/luasrc/view/syncdial/redial_button.htm
@@ -2,9 +2,9 @@
 <script type="text/javascript" src="<%=resource%>/cbi.js"></script>
 <script type="text/javascript">//<![CDATA[
 
-	function do_macvlan_rediag()
+	function do_macvlan_redial()
 	{
-		XHR.get('<%=luci.dispatcher.build_url("admin", "network", "macvlan_rediag")%>', null,
+		XHR.get('<%=luci.dispatcher.build_url("admin", "network", "macvlan_redial")%>', null,
 			function(x, st)
 			{
 				window.location.href='<%=luci.dispatcher.build_url("admin", "network", "network")%>';
@@ -13,5 +13,5 @@
 	}
 //]]></script>
 
-<input class="cbi-button cbi-input-save" id="cbi-macvlan_rediag-<%=section%>-action" type="button" value="<%:rediag%>" onclick="do_macvlan_rediag()" />
+<input class="cbi-button cbi-input-save" id="cbi-macvlan_rediag-<%=section%>-action" type="button" value="重新拨号" onclick="do_macvlan_redial()" />
 <%+cbi/valuefooter%>
