+++ applications/luci-app-syncdial/root/etc/config/syncdial
@@ -1,0 +1,6 @@
+config syncdial 'config'
+	option 'enabled'	'0'
+	option 'wannum'		'2'
+	option 'dialchk'	'0'
+	option 'dialnum'	'1'
+	option 'force_redial'	'0'
