+++ applications/luci-app-syncdial/root/etc/hotplug.d/iface/01-dialcheck
@@ -1,0 +1,4 @@
+#!/bin/sh
+[ "$ACTION" = "ifdown" ] && pppconnectcheck &
+
+
