+++ applications/luci-app-syncdial/root/etc/uci-defaults/luci-syncdial
@@ -1,0 +1,12 @@
+#!/bin/sh
+touch /etc/config/syncdial
+
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@syncdial[-1]
+	add ucitrack syncdial
+	set ucitrack.@syncdial[-1].exec='/bin/genwancfg'
+	commit ucitrack
+EOF
+
+rm -f /tmp/luci-indexcache
+exit 0
